Snap3D/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
rembg[cpu]
Pillow
torch
torchvision
trimesh>=4.4.0
numpy
scipy
omegaconf==2.3.0
einops==0.7.0
transformers==4.35.0
huggingface-hub==0.19.4
xatlas==0.0.9
moderngl==5.10.0
imageio[ffmpeg]
qrcode[pil]
git+https://github.com/tatsy/torchmcubes.git
